refactor(util, generator): generator 패키지 분리 및 InputSplitter 유틸성 클래스화
- generator 관련 파일을 util 패키지에서 분리하여 독립된 패키지로 이동
- InputSplitter를 유틸리티 클래스로 변경하고 static 메서드로 구조 개선
- InputParser 인터페이스 삭제
