+++ src/main/java/racingcar/Application.java
@@ -1,14 +1,12 @@
 package racingcar;
 
 import racingcar.controller.CarController;
+import racingcar.generator.Generator;
+import racingcar.generator.IntegerGenerator;
 import racingcar.model.CarNames;
 import racingcar.service.RacingService;
 import racingcar.strategy.CarMoveStrategy;
 import racingcar.strategy.MoveStrategy;
-import racingcar.util.generator.Generator;
-import racingcar.util.generator.IntegerGenerator;
-import racingcar.util.parser.InputParser;
-import racingcar.util.parser.InputSplitter;
 import racingcar.validator.CarCountValidator;
 import racingcar.validator.CarNameValidator;
 import racingcar.validator.RoundValidator;
@@ -20,8 +18,6 @@ import racingcar.view.ViewFacade;
 
 public class Application {
     public static void main(String[] args) {
-        InputParser inputParser = new InputSplitter();
-
         MoveStrategy moveStrategy = new CarMoveStrategy();
         Generator<Integer> generator = new IntegerGenerator();
         RacingService racingService = new RacingService(moveStrategy, generator);
@@ -35,7 +31,7 @@ public class Application {
         OutputView outputView = new OutputView();
         ViewFacade viewFacade = new ViewFacade(inputView, outputView);
 
-        CarController carController = new CarController(racingService, validatorFacade, inputParser, viewFacade);
+        CarController carController = new CarController(racingService, validatorFacade, viewFacade);
 
         carController.play();
     }
