+++ src/main/java/racingcar/generator/Generator.java
@@ -1,4 +1,4 @@
-package racingcar.util.generator;
+package racingcar.generator;
 
 public interface Generator<T> {
     T generate();
