--- src/main/java/racingcar/util/parser/InputParser.java
@@ -1,8 +1,0 @@
-package racingcar.util.parser;
-
-
-import racingcar.model.CarNames;
-
-public interface InputParser {
-    CarNames parse(String input);
-}
